+++ app/Http/Controllers/Tenant/FornecedoresController.php
@@ -14,6 +14,8 @@ class FornecedoresController extends Controller
 {
     public function index(): View
     {
+        $this->authorize('viewAny', Fornecedor::class);
+
         $fornecedores = Fornecedor::orderBy('razao_social')->paginate(25);
 
         return view('tenant.fornecedores.index', compact('fornecedores'));
@@ -21,11 +23,15 @@ class FornecedoresController extends Controller
 
     public function create(): View
     {
+        $this->authorize('create', Fornecedor::class);
+
         return view('tenant.fornecedores.create');
     }
 
     public function store(StoreFornecedorRequest $request): RedirectResponse
     {
+        $this->authorize('create', Fornecedor::class);
+
         $data = $request->validated();
         $data['cnpj'] = FornecedorService::formatarCnpj($data['cnpj']);
 
@@ -37,11 +43,15 @@ class FornecedoresController extends Controller
 
     public function edit(Fornecedor $fornecedor): View
     {
+        $this->authorize('update', $fornecedor);
+
         return view('tenant.fornecedores.edit', compact('fornecedor'));
     }
 
     public function update(UpdateFornecedorRequest $request, Fornecedor $fornecedor): RedirectResponse
     {
+        $this->authorize('update', $fornecedor);
+
         $data = $request->validated();
         $data['cnpj'] = FornecedorService::formatarCnpj($data['cnpj']);
 
@@ -53,6 +63,8 @@ class FornecedoresController extends Controller
 
     public function destroy(Fornecedor $fornecedor): RedirectResponse
     {
+        $this->authorize('delete', $fornecedor);
+
         $fornecedor->delete();
 
         return redirect()->route('tenant.fornecedores.index')
