+++ app/Http/Controllers/Tenant/SecretariasController.php
@@ -13,6 +13,8 @@ class SecretariasController extends Controller
 {
     public function index(): View
     {
+        $this->authorize('viewAny', Secretaria::class);
+
         $secretarias = Secretaria::orderBy('nome')->paginate(25);
 
         return view('tenant.secretarias.index', compact('secretarias'));
@@ -20,11 +22,15 @@ class SecretariasController extends Controller
 
     public function create(): View
     {
+        $this->authorize('create', Secretaria::class);
+
         return view('tenant.secretarias.create');
     }
 
     public function store(StoreSecretariaRequest $request): RedirectResponse
     {
+        $this->authorize('create', Secretaria::class);
+
         Secretaria::create($request->validated());
 
         return redirect()->route('tenant.secretarias.index')
@@ -33,11 +39,15 @@ class SecretariasController extends Controller
 
     public function edit(Secretaria $secretaria): View
     {
+        $this->authorize('update', $secretaria);
+
         return view('tenant.secretarias.edit', compact('secretaria'));
     }
 
     public function update(UpdateSecretariaRequest $request, Secretaria $secretaria): RedirectResponse
     {
+        $this->authorize('update', $secretaria);
+
         $secretaria->update($request->validated());
 
         return redirect()->route('tenant.secretarias.index')
@@ -46,6 +56,8 @@ class SecretariasController extends Controller
 
     public function destroy(Secretaria $secretaria): RedirectResponse
     {
+        $this->authorize('delete', $secretaria);
+
         $secretaria->delete();
 
         return redirect()->route('tenant.secretarias.index')
